refactored out the color, vector2D and keyboardEnums into separate js files for readabilty
