+++ src/index.html
@@ -4,13 +4,12 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Pathtracer</title>
-    <script type="text/javascript" src="index.js"></script>
+    <script type="module" src="index.js"></script>
 </head>
 
 <body>
     <p id="result"></p>
     <canvas id="viewport" width="320" height="240"></canvas>
-
     <div><input id="framelimiter" type="number" value="30" /> </div>
 </body>
