+++ src/modules/color.js
@@ -1,0 +1,12 @@
+class Color {
+    constructor(r, g, b, a) {
+        this.r = r;
+        this.g = g;
+        this.b = b;
+        this.a = a;
+    }
+}
+
+export {
+    Color
+};
